Doesn't shake as much.
