+++ odo.js
@@ -34,3 +34,27 @@ exports.streamOdo = async function streamOdo(sprkp) {
         console.log('  y: ' + global.posy + data.yOdometer.units);*/
     });
 }
+
+exports.reached_target = function reached_target(target) {
+// Environmental parameters
+const outerDist = 45;
+const innerDist = 35;
+const slack = 3;
+const target_width = 2 * slack;
+let start_a = target - slack;
+let end_a = target + slack;
+
+// Positional
+let start_dist = pa - start_a;
+let pos_rel_a = 0;
+if (start_dist > 0) {
+    pos_rel_a = start_dist;
+} else {
+    pos_rel_a = start_dist + 360;
+}
+
+let dist = Math.sqrt(px * px + py * py);
+let valid_dist = innerDist < dist < outerDist;
+
+return (0 <= pos_rel_a && pos_rel_a <= target_width && valid_dist)
+}
